+++ fleetpings/helper/eve_images.py
@@ -1,5 +1,5 @@
 """
-eve images
+Eve images
 """
 
 # Alliance Auth