+++ fleetpings/tests/test_auth_hooks.py
@@ -1,0 +1,73 @@
+"""
+Test auth_hooks
+"""
+
+# Standard Library
+from http import HTTPStatus
+
+# Django
+from django.test import TestCase
+from django.urls import reverse
+
+# AA Fleet Pings
+from fleetpings.tests.utils import create_fake_user
+
+
+class TestHooks(TestCase):
+    """
+    Test the app hook into allianceauth
+    """
+
+    @classmethod
+    def setUpClass(cls) -> None:
+        """
+        Set up groups and users
+        """
+
+        super().setUpClass()
+
+        # User cannot access
+        cls.user_1001 = create_fake_user(1001, "Peter Parker")
+
+        # User can access
+        cls.user_1002 = create_fake_user(
+            1002, "Bruce Wayne", permissions=["fleetpings.basic_access"]
+        )
+
+        cls.html_menu = f"""
+            <li>
+                <a class href="{reverse('fleetpings:index')}">
+                    <i class="far fa-bell fa-fw"></i>
+                    Fleet Pings
+                </a>
+            </li>
+        """
+
+    def test_render_hook_success(self):
+        """
+        Test should show the link to the app in the navigation to user with access
+        :return:
+        :rtype:
+        """
+
+        self.client.force_login(self.user_1002)
+
+        response = self.client.get(reverse("authentication:dashboard"))
+
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        self.assertContains(response, self.html_menu, html=True)
+
+    def test_render_hook_fail(self):
+        """
+        Test should not show the link to the app in the
+        navigation to user without access
+        :return:
+        :rtype:
+        """
+
+        self.client.force_login(self.user_1001)
+
+        response = self.client.get(reverse("authentication:dashboard"))
+
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        self.assertNotContains(response, self.html_menu, html=True)