+++ testauth/__init__.py
@@ -1,1 +1,5 @@
+"""
+Initialize the test suite
+"""
+
 from .celery import app as celery_app  # noqa: F401